Cycle fixes: extract library constants
