+++ src/ui/components/UploadScreen/UploadScreen.tsx
@@ -7,7 +7,8 @@ import { Recording, ExperimentalUserSettings } from "ui/types";
 import LoadingScreen from "../shared/LoadingScreen";
 import { useGetRecordingId } from "ui/hooks/recordings";
 import { trackEvent } from "ui/utils/telemetry";
-import Sharing, { MY_LIBRARY } from "./Sharing";
+import { MY_LIBRARY } from "./libraryConstants";
+import Sharing from "./Sharing";
 import { Privacy, ToggleShowPrivacyButton } from "./Privacy";
 import { UploadRecordingTrialEnd } from "./UploadRecordingTrialEnd";
 import { startUploadWaitTracking } from "ui/utils/mixpanel";
