+++ src/ui/components/UploadScreen/libraryConstants.ts
@@ -1,0 +1,2 @@
+export const MY_LIBRARY = "Your Library";
+export const personalWorkspace = { id: MY_LIBRARY, name: MY_LIBRARY };
